get.data.dir {fadnUtils}	R Documentation

Gets the data.dir

Description

data.dir is the folder where data is stored r package will create two subfolders: csv = location to store the csv files of th DG-AGRI (fadn.raw.csv) rds = location to store rds files (fadn.raw.rds, fadn.str.rds, etc.)

Usage

get.data.dir()

Value

the value of option("fadnUtils.data.dir")

[Package fadnUtils version 1.0.2 Index]