grep.columns.in.raw.rds {fadnUtils}	R Documentation

Grep a pattern into a raw.rds column names

Description

Useful for the case where one want to look if there are certain columns present or missing

Usage

grep.columns.in.raw.rds(pattern, countries = c("all"), years = c("all"))

Arguments

pattern
a grep-like character pattern. This parameter is passed as is to the grep function

countries
a character vector with all the 3-letter codes of the selected countries, e.g. c("ELL", "ESP"). If "all" is included, all available countries are loaded

years
a numeric vector with the years selected. If "all" is included, all available years are loa

show
if TRUE, the columsn are printed

Value

Prints the columns and returns them invisibly

[Package fadnUtils version 1.0.2 Index]