raw_str_map.merge {fadnUtils}	R Documentation

Merges two raw_str_map files and returns either a list or a file

Description

All entries in the new.raw_str_map file replace those on the source.raw_str_map file

Usage

raw_str_map.merge(
  source.raw_str_map.file = NULL,
  new.raw_str_map.file = NULL,
  return.file = F
)

Arguments

source.raw_str_map.file
the filename of the source raw_str_map. It must be relative the raw_str_maps of the current data.dir

new.raw_str_map.file
the filename of the mask raw_str_map. It will replace any entries of the source file. It must be relative the raw_str_maps of the current data.dir

return.file
If set to T, a temporary full file path that contains the merge is returned. Otherwise a list with the contents of the merge is returned

Details

Both files must be relative to the current data.dir/raw_str_maps

Value

FALSE in case of problem / if return.file=T, the temporary full path of a file that contains the merged result in json / A list with the contents of the merge if return.file=F

[Package fadnUtils version 1.0.2 Index]