show.data.dir.contents {fadnUtils}	R Documentation

Show the contents of data.dir

Description

Show the contents of data.dir

Usage

show.data.dir.contents(data.dir = NULL)

Arguments

data.dir
a specific directory to show contents, otherwise it will read the fadnUtils.data.dir

return.list
if T, returns a list, otherwise print the results

Value

returns a list containing DT of fadn.raw.rds (Country-Year) and of fadn.str.rds (country-Year)

[Package fadnUtils version 1.0.2 Index]