write.excel {fadnUtils}	R Documentation

Utility to copy data to clipboard for pasting to Excel

Description

Utility to copy data to clipboard for pasting to Excel

Usage

write.excel(d, getRownames = F, ...)

Arguments

d
the data to copy

getRownames
set to T to opy also row.names

...
any other parameter for passing to write.table

Value

nothing

Examples

write.excel(d);

[Package fadnUtils version 1.0.2 Index]